pp/schema_registry: extract proto test util functions
Extract protobuf testing utility function into their own library
