+++ src/v/pandaproxy/schema_registry/test/compatibility_protobuf.cc
@@ -16,6 +16,7 @@
 #include "pandaproxy/schema_registry/protobuf.h"
 #include "pandaproxy/schema_registry/sharded_store.h"
 #include "pandaproxy/schema_registry/test/compatibility_common.h"
+#include "pandaproxy/schema_registry/test/protobuf_utils.h"
 #include "pandaproxy/schema_registry/types.h"
 
 #include <seastar/testing/thread_test_case.hh>
@@ -28,48 +29,15 @@
 
 namespace pp = pandaproxy;
 namespace pps = pp::schema_registry;
+namespace ppstu = pp::schema_registry::test_utils;
 
 namespace {
-
-struct simple_sharded_store {
-    explicit simple_sharded_store()
-      : store{} {
-        store.start(pps::is_mutable::yes, ss::default_smp_service_group())
-          .get();
-    }
-    ~simple_sharded_store() { store.stop().get(); }
-    simple_sharded_store(const simple_sharded_store&) = delete;
-    simple_sharded_store(simple_sharded_store&&) = delete;
-    simple_sharded_store& operator=(const simple_sharded_store&) = delete;
-    simple_sharded_store& operator=(simple_sharded_store&&) = delete;
-
-    pps::schema_id
-    insert(const pps::subject_schema& schema, pps::schema_version version) {
-        const auto id = next_id++;
-        store
-          .upsert(
-            pps::seq_marker{
-              std::nullopt,
-              std::nullopt,
-              version,
-              pps::seq_marker_key_type::schema},
-            schema.share(),
-            id,
-            version,
-            pps::is_deleted::no)
-          .get();
-        return id;
-    }
-
-    pps::schema_id next_id{1};
-    pps::sharded_store store;
-};
 
 bool check_compatible(
   pps::compatibility_level lvl,
   std::string_view reader,
   std::string_view writer) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
     store.store.set_compatibility(lvl).get();
     store.insert(
       pandaproxy::schema_registry::subject_schema{
@@ -101,7 +69,7 @@ pps::compatibility_result check_compatible_verbose(
 } // namespace
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_simple) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto schema1 = pps::subject_schema{pps::subject{"simple"}, simple.share()};
     store.insert(schema1.share(), pps::schema_version{1});
@@ -112,7 +80,7 @@ SEASTAR_THREAD_TEST_CASE(test_protobuf_simple) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_nested) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto schema1 = pps::subject_schema{pps::subject{"nested"}, nested.share()};
     store.insert(schema1.share(), pps::schema_version{1});
@@ -125,7 +93,7 @@ SEASTAR_THREAD_TEST_CASE(test_protobuf_nested) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_imported_failure) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     // imported depends on simple, which han't been inserted
     auto schema1 = pps::subject_schema{
@@ -142,7 +110,7 @@ SEASTAR_THREAD_TEST_CASE(test_protobuf_imported_failure) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_imported_not_referenced) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto schema1 = pps::subject_schema{pps::subject{"simple"}, simple.share()};
     auto schema2 = pps::subject_schema{
@@ -162,7 +130,7 @@ SEASTAR_THREAD_TEST_CASE(test_protobuf_imported_not_referenced) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_referenced) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto schema1 = pps::subject_schema{
       pps::subject{"simple.proto"}, simple.share()};
@@ -187,7 +155,7 @@ SEASTAR_THREAD_TEST_CASE(test_protobuf_referenced) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_recursive_reference) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto schema1 = pps::subject_schema{
       pps::subject{"simple.proto"}, simple.share()};
@@ -212,7 +180,7 @@ SEASTAR_THREAD_TEST_CASE(test_protobuf_recursive_reference) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_binary_protobuf) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     BOOST_REQUIRE_NO_THROW(
       store.store
@@ -225,7 +193,7 @@ SEASTAR_THREAD_TEST_CASE(test_binary_protobuf) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_invalid_binary_protobuf) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto broken_base64_raw_proto = base64_raw_proto.substr(1);
 
@@ -250,7 +218,7 @@ SEASTAR_THREAD_TEST_CASE(test_invalid_binary_protobuf) {
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_well_known) {
-    simple_sharded_store store;
+    ppstu::simple_sharded_store store;
 
     auto schema = pps::subject_schema{
       pps::subject{"test_auto_well_known"},
@@ -501,27 +469,6 @@ SEASTAR_THREAD_TEST_CASE(
       pps::compatibility_level::full_transitive, recursive, recursive));
 }
 
-auto sanitize(
-  std::string_view raw_proto, pps::normalize norm = pps::normalize::no) {
-    simple_sharded_store s;
-    iobuf buf = pps::make_canonical_protobuf_schema(
-                  s.store,
-                  pps::subject_schema{
-                    pps::subject{"foo"},
-                    pps::schema_definition{
-                      raw_proto, pps::schema_type::protobuf}},
-                  norm)
-                  .get()
-                  .def()
-                  .raw()();
-    iobuf_parser parser{std::move(buf)};
-    return parser.read_string(parser.bytes_left());
-}
-
-auto normalize(std::string_view raw_proto) {
-    return sanitize(raw_proto, pps::normalize::yes);
-}
-
 constexpr auto foobar_proto = R"(syntax = "proto3";
 package foo;
 
@@ -534,7 +481,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_strip_comments_and_newlines) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(
+      ppstu::sanitize(R"(
 
 /* comment */
 
@@ -562,7 +509,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_ordering_no_newlines) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(syntax = "proto3";
+      ppstu::sanitize(R"(syntax = "proto3";
 import "google/protobuf/timestamp.proto";
 package foo;
 message Bar {
@@ -574,7 +521,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_ordering_more_newlines) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(
+      ppstu::sanitize(R"(
 
 syntax = "proto3";
 
@@ -869,8 +816,8 @@ extend .google.protobuf.ServiceOptions {
 }
 
 )";
-    BOOST_CHECK_EQUAL(sanitize(schema), sanitized);
-    BOOST_CHECK_EQUAL(normalize(schema), normalized);
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), sanitized);
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), normalized);
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_normalize_nested_custom_options) {
@@ -987,8 +934,8 @@ extend .google.protobuf.MessageOptions {
 }
 
 )";
-    BOOST_CHECK_EQUAL(sanitize(schema), sanitized);
-    BOOST_CHECK_EQUAL(normalize(schema), normalized);
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), sanitized);
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), normalized);
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_normalize_message_custom_options) {
@@ -1059,8 +1006,8 @@ extend .google.protobuf.EnumValueOptions {
 
 )";
 
-    BOOST_REQUIRE_EQUAL(sanitize(schema), sanitized);
-    BOOST_CHECK_EQUAL(normalize(schema), normalized);
+    BOOST_REQUIRE_EQUAL(ppstu::sanitize(schema), sanitized);
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), normalized);
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_normalize_extension_ranges) {
@@ -1195,13 +1142,13 @@ extend .google.protobuf.ExtensionRangeOptions {
 
 )";
 
-    BOOST_CHECK_EQUAL(sanitize(schema), sanitized);
-    BOOST_CHECK_EQUAL(normalize(schema), normalized);
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), sanitized);
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), normalized);
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_no_syntax) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(
+      ppstu::sanitize(R"(
 package foo;
 
 import "google/protobuf/timestamp.proto";
@@ -1223,7 +1170,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_no_package) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(syntax = "proto3";
+      ppstu::sanitize(R"(syntax = "proto3";
 
 import "google/protobuf/timestamp.proto";
 
@@ -1243,7 +1190,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_no_syntax_package) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(
+      ppstu::sanitize(R"(
 
 import "google/protobuf/timestamp.proto";
 
@@ -1263,7 +1210,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_no_imports) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(syntax = "proto3";
+      ppstu::sanitize(R"(syntax = "proto3";
 package foo;
 
 message Bar {
@@ -1283,7 +1230,7 @@ message Bar {
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_sanitize_multiple_imports) {
     BOOST_REQUIRE_EQUAL(
-      sanitize(R"(syntax = "proto3";
+      ppstu::sanitize(R"(syntax = "proto3";
 
 package foo;
 
@@ -1321,7 +1268,7 @@ import weak "google/protobuf/any.proto";
 import "google/protobuf/api.proto";
 )";
 
-    BOOST_CHECK_EQUAL(sanitize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), (R"(syntax = "proto3";
 package foo;
 
 import "google/protobuf/timestamp.proto";
@@ -1331,7 +1278,7 @@ import public "google/protobuf/duration.proto";
 
 
 )"));
-    BOOST_CHECK_EQUAL(normalize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), (R"(syntax = "proto3";
 package foo;
 
 import "google/protobuf/api.proto";
@@ -1359,7 +1306,7 @@ message HasGoogleMap {
 }
 )";
 
-    BOOST_CHECK_EQUAL(sanitize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), (R"(syntax = "proto3";
 package foo;
 
 import "google/protobuf/struct.proto";
@@ -1377,7 +1324,7 @@ message HasGoogleMap {
   map<string, .google.protobuf.Value> map_string_value = 1;
 }
 )"));
-    BOOST_CHECK_EQUAL(normalize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), (R"(syntax = "proto3";
 package foo;
 
 import "google/protobuf/any.proto";
@@ -1423,7 +1370,7 @@ message HasMap {
 }
 )";
 
-    BOOST_CHECK_EQUAL(sanitize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), (R"(syntax = "proto3";
 
 import "google/protobuf/timestamp.proto";
 message Value {
@@ -1436,7 +1383,7 @@ message HasMap {
 }
 
 )"));
-    BOOST_CHECK_EQUAL(normalize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), (R"(syntax = "proto3";
 
 import "google/protobuf/timestamp.proto";
 message Value {
@@ -1471,7 +1418,7 @@ message SearchResponse {
   }
 })";
 
-    BOOST_CHECK_EQUAL(sanitize(schema), (R"(syntax = "proto2";
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), (R"(syntax = "proto2";
 
 message SearchResponse {
   repeated group Result = 1 {
@@ -1492,7 +1439,7 @@ message SearchResponse {
 }
 
 )"));
-    BOOST_CHECK_EQUAL(normalize(schema), (R"(syntax = "proto2";
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), (R"(syntax = "proto2";
 
 message SearchResponse {
   repeated group Result = 1 {
@@ -1544,7 +1491,7 @@ message WithOneOf {
 }
 
 )";
-    BOOST_CHECK_EQUAL(sanitize(schema), expected_sanitized);
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), expected_sanitized);
     auto expected_normalized = R"(syntax = "proto3";
 
 package foo;
@@ -1560,7 +1507,7 @@ message WithOneOf {
 }
 
 )";
-    BOOST_CHECK_EQUAL(normalize(schema), expected_normalized);
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), expected_normalized);
 }
 
 SEASTAR_THREAD_TEST_CASE(test_protobuf_normalize) {
@@ -1664,7 +1611,7 @@ service FooService {
   rpc Foo(Bar) returns (Baz);
 })";
 
-    BOOST_CHECK_EQUAL(sanitize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::sanitize(schema), (R"(syntax = "proto3";
 package foo;
 
 import "google/protobuf/timestamp.proto";
@@ -1741,7 +1688,7 @@ service FooService {
   rpc Foo(.foo.Bar) returns (.foo.Baz);
 }
 )"));
-    BOOST_CHECK_EQUAL(normalize(schema), (R"(syntax = "proto3";
+    BOOST_CHECK_EQUAL(ppstu::normalize(schema), (R"(syntax = "proto3";
 package foo;
 
 import "google/protobuf/any.proto";
